Add min length property and computed explanatory prompt to the component
modified first expectation to just match the original prompt is contained
in the rendered component.

Added new expectation to ensure the computed full prompt has both the prompt 
and the value of minimum length appended to it.

There's a bit of duplication with wrapper definition, but trying to move it
into a beforeEach() method failed, so need to figure out how to abstract.

Also it's good idea to load the wrapper with only the required data, so I didn't add the minLength property to the wrapper in the first expectation as it's not needed in that test.