Fix a test that was passing for the wrong reason
The test that detects if a username is too short was passing,
but the expectation passed because of the empty string username passed to the factory,
not because it detected that the hand-crafted shortOne username was too short.

This is because when using setData to alter the state of shallow mounted wrapper,
I have to wait for the next tick in the change update life cycle.
Because I didn't use nextTick(), the username picked up by the wrapper call in the expectation
was the original empty string one.

Fixed it by wrapping the expectation in a Promise returned by nextTick().
Found this issue when about to implement a refinement where the error is
not displayed if the username is empty (that will be in the next commit)

More info here:
* https://vue-test-utils.vuejs.org/guides/#using-nexttick-2