Refine UX: error is accounted for but not displayed if username is empty
because it will irritate the user to get the error on the onset.

Implementing this exposed the issue fixed in the previous commit where the test for
too short username was picking up the wrong state of the component
(testing against the initial value of username which was empty string instead
of the updated value after setting up the test username).

In the test here we use the Truthy test for the computed property 'error'
because it's not true as in the boolean type, it's a non-empty string which
count as "truthy".

For testing the display of the .error div, we use the absolute boolean test,
because exists() returns either true or false as in the boolean types.