Add page size display to Page Info component
It's in the 'computed:', not 'methods:' section of the component properties
because it's fine (and desirable) for this value to be cached.

We use Jasmine's Spy class for test doubles to mock the document.documentElement
so we can control the page size in test to verify with different values.
Because document.documentElement is not a method we can mock it, however
document.querySelector('html') points to the same element, so that's
the the function that we can stub.