Setup coverage more correctly with karma-coverage, babel-plugin-istanbul
Now there is a proper babel config file (babel.config.js) where the istanbul plugin is mentioned
for the test environment only. Which means that I need the cross-env package to be able to specify to karma command
which babel environment to pass.

Now that babel has a config file the loader rule in webpack config file can be simpler

The coverage report is still odd (I think code instrumentation may not happen properly),
but at least the configuration is in a better state than before